Run migrations and seed the database as a postbuild action for review apps
Runs dbmigrate up to migrate the database up to the current schema.
Loads the seeds into the review app database

Note that postdeploy is run only once, after the app has been created and
deployed for the first time. It is not run when subsequent changes are deployed.

See https://devcenter.heroku.com/articles/github-integration-review-apps#the-postdeploy-script

Co-authored-by: Joshua-Luke Bevan <joshua.bevan@madetech.com>