Remove support link from visit completed page for now
This is specific to a trust. As we are moving towards multiple trusts, we need to gear this or find a good support source which is suitable for all trusts. Atm, we don't have a way to get each support page for each trust so for now we'll have to remove it.