Close the database connection after all tests have ran
Previously we were running them after each test. This causes issues when we
have multiple tests in the same file as the connection will be closed before
the next test can run.

By moving the teardown to a seperate file, developers do no need to include
it in any new contract tests

Although we do have a testSetup.js, importing the AppContainer in that file
resulted in errors in the GovNotify contract test