Use moment().unix to parse the token expiry (Unix timestamp)
The token expiry is stored as a unix timestamp (seconds since epoch).
Previously we were multiplying this by 1000 to convert it to milliseconds so
that it could be passed in to Date. This change replaces that with the unix function
so that there is no confusion about why we are multiplying