Hide the hospital name column if the ward visit totals are displayed
The hospital name will eventually be removed. But while it is still appearing
on the TrustAdmin dashboard we want to show the Hospital Name column when the
table is rendered there. If we are showing ward visit totals then we're assuming
we're on the view hospital page and can hide the hospital name column