Add pageLoadTimeout to places in Cypress tests where API is doing a lot
Including booking a visit and starting a visit to avoid flakey tests. These
expectations need more time as the APIs for these are also sending
messages and emails or calling other APIs. Therefore, more time is
needed for these actions to complete before the next page is shown.

We tried to rectify this using `cy.url` but upon looking at the docs,
this has the same timeout as `cy.get` so an optional timeout has been
set to the Cypress' pageLoadTimeout as if `cy.visit` was used to view
another page.