Move trust creation to a setupTrust fixture
This change introduces a new fixture, 'setupTrust', which allows us to quickly
setup a trust in a test without having to worry about making sure all the values
are set correctly.

If we need a specific value for a test, one can be passed in to the fixture to
override the defaults.

This allows us make changes to trust creation without having to change all the tests
(e.g. adding a new required field)