Fix sending an updated notification for editing a visit with no changes
This was because when determining the notification type we were
comparing a string and an object for the call time which was not
resolving to true.

When a request is made the call time is received as a string and
therefore needs to be converted to a object to match the retrieved call
time from the database.