Update visit retrieving usecases to include complete visits that have not been cleared
We eventually want to be able to mark visits at complete but still show them
in the list screen. To do that we need to be able to differentiate between
visits that have been cleared and which have not.

This change updates the visit retrieving usecases to check for complete visits
where pii_cleared_at IS NULL, which signifies that they are safe to display
to the user