Extract manual fork sync action to its own workflow
The fork sync workflow skips if the repo doesn't match the ones referenced in
the workflow. If you attempted to manually run that workflow on another repo, it
would be skipped because it was checking the same conditional.

This change extracts the manual sync to its own workflow so that you can run it
even if your repo is not in the list of automatically synced repos