Fix sending "null" to book a visit form from confirmation page
The end to end test for book a visit was failing because it failed
validation after first passing validation, going to the check your
answers page, clicking a Change link and then submitting the form.

This is because we pass the input values between the form and check your
answers page using query params. When we passed back from the check your
answers page, for some reason it passed back a string of "null" for
contact number. As a result, the form prepopulated the contact number
input with the value "null" which failed the phone number validation.