stop automatic redirect to `ward/lvisits`
the redirect happens based on the token which has been set on the cookie.
The token is generated based on the login code provided. If the login
code is the same for a ward as it is for the admin then the
cookie will be set for both. This is because the ward login does not
require a password its code to be verified which is required to generate
a token. The same applies for when the trust admin code is the same as
the ward code. So in both cases a token for a ward is also set which
permits a redirect to the ward visits page. To prevent this a guard
clause is set to check if a verified code exists for the trust admin or
the admin before generating a token for the ward login.