Attempt to update next & React (#980)
* Attempt to update next & React

* Work on refactoring the CSS in the project to align with the requirements of NextJS 11.

As of Next 11 the previously-used SASS config is considered deprecated and should not be used, in favour of using the pre-configured SASS that ships with NextJS. Part of this config disallows the use of global stylesheets anywhere other than _app.js

This commit refactors the SASS in the project to:
* Move global CSS into a single file
* Refactor existing CSS into CSS modules
* Tidy up imports so that they only import the things they need

* Removed leftover commented-out code from testing, removed @zeit/next-sass dependency

* Regenerate package-lock with the correct node version

* Disable E2E Tests.

As the End-to-end tests are showing significant flake, we are going to temporarily disable them so as to be able to take stock of the state of the rest of the pipeline. Once this has been done, we will re-enable the E2E tests.